Move :lastfm module to :network:lastfm
