+++ network/lastfm/src/main/AndroidManifest.xml
