+++ network/lastfm/src/main/java/de/schnettler/lastfm/Utils.kt
