+++ network/lastfm/src/main/java/de/schnettler/lastfm/annotation/retrofit/AuthorizedLastfmRetrofitClient.kt
