+++ network/lastfm/src/main/java/de/schnettler/lastfm/api/lastfm/DetailService.kt
