+++ network/lastfm/src/main/java/de/schnettler/lastfm/api/lastfm/SearchService.kt
