+++ network/lastfm/src/main/java/de/schnettler/lastfm/models/BaseInfoDto.kt
