+++ network/spotify/src/main/java/de/schnettler/scrobbler/network/spotify/SpotifyAuthProvider.kt
@@ -3,4 +3,5 @@ package de.schnettler.scrobbler.network.spotify
 
 interface SpotifyAuthProvider {
     suspend fun getAuthToken(): String
+    suspend fun refreshToken(): String
 }
